First move of a utility common to both the windows and cygwin modules.
Although I considered adding a new module for this purpose, for now it
suffices to use the _common module for this (albeit a bit hackishly).

In this case the C wrapper for this function makes sense to put in
arch/windows.
